% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{estp_gdp}
\alias{estp_gdp}
\title{Small p-values estimation with GDP mle estimators}
\usage{
estp_gdp(Nexc = 250, list_stats, sub_out)
}
\arguments{
\item{Nexc}{Exceedances threshold (default 250)}

\item{list_stats}{List of statistics from simulations}

\item{sub_out}{Result from get_sum_stats function}
}
\value{
An estimated p-value
}
\description{
Small p-values estimation with GDP mle estimators
}
\examples{
list_stats <- c(0.25, 1.69, 1.86, 2.3, 0.01)
p_value <- estp_gdp(250, list_stats, sub_out)
}
