% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{get_all_genes}
\alias{get_all_genes}
\title{Get all genes}
\usage{
get_all_genes(files)
}
\arguments{
\item{files}{a list of all TWAS results files}
}
\value{
a dataframe with all gene symbols (IDs)
}
\description{
Get all genes
}
\examples{
gene_list <= get_all_genes(files)
}
