% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{get_sum_stats}
\alias{get_sum_stats}
\title{Get the test statistics}
\usage{
get_sum_stats(Z)
}
\arguments{
\item{Z}{z-values transformed from TWAS p-values}
}
\value{
a list with test statistics (MaxS), corresponding selected subset (idx), and all possible statistics
}
\description{
Get the test statistics
}
\examples{
z <- 1:3
names(z) <- c("a", "b", "c")
test_stats <- get_sum_stats(z)
}
