% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mhp_cstwas}
\alias{mhp_cstwas}
\title{Manhattan Plot For the CSTWAS Resutls}
\usage{
mhp_cstwas(
  meta_data,
  cstwas_res,
  anot_index = 15,
  pts_size = 2,
  ceiling_ctf = NULL,
  floor_ctf = 0,
  anno_gene = NULL,
  path = NULL
)
}
\arguments{
\item{meta_data}{meta_data from run_CSTWAS results.}

\item{cstwas_res}{cstwas_res from run_CSTWAS results.}

\item{anot_index}{An integer indicating how significant results are to be annotated. (-log10(TWAS.P) > anot_index) This parameter will be ignored if anno_gene is not NULL.}

\item{pts_size}{An integer indicating the point size.}

\item{ceiling_ctf}{An integer indicating how significant results are to be cut by the ceiling. (-log10(TWAS.P) > ceiling_ctf). If is NULL, it will automatically adjust based on the data.}

\item{floor_ctf}{An integer indicating how insignificant results are to be cut by the floor (-log10(TWAS.P) < floor_ctf). Default 0.}

\item{anno_gene}{A list of genes that need to be annotated.}

\item{path}{Path for saving the plot.}
}
\value{
A Manhattan plot
}
\description{
Manhattan Plot For the CSTWAS Resutls
}
\examples{
mhp_cstwas(test$meta_data, test$cstwas_res)
}
