% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mhp_sctwas}
\alias{mhp_sctwas}
\title{Manhattan Plot For the Subset-based Cross-tissue TWAS Resutls}
\usage{
mhp_sctwas(
  meta_data,
  sctwas_res,
  anot_index = 15,
  ceiling_ctf = 30,
  anno_gene = NULL,
  path = NULL
)
}
\arguments{
\item{meta_data}{meta_data from run_SCTWAS results.}

\item{sctwas_res}{sctwas_res from run_SCTWAS results.}

\item{anot_index}{An integer indicating how significant results are to be annotated. (-log10(TWAS.P) > anot_index) This parameter will be ignored if anno_gene is not NULL.}

\item{ceiling_ctf}{An integer indicating how significant results are to be cut by the ceiling. (-log10(TWAS.P) > ceiling_ctf)}

\item{anno_gene}{A list of genes that need to be annotated.}

\item{path}{Path for saving the plot.}
}
\value{
A Manhattan plot
}
\description{
Manhattan Plot For the Subset-based Cross-tissue TWAS Resutls
}
\examples{
mhp_sctwas(test$meta_data, test$sctwas_res)
}
