% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mhp_twas}
\alias{mhp_twas}
\title{Manhattan Plot For TWAS Results}
\usage{
mhp_twas(
  meta_data,
  anot_index = 15,
  ceiling_ctf = 30,
  pts_size = 3.6,
  anno_gene = NULL,
  path = NULL
)
}
\arguments{
\item{meta_data}{meta_data from run_SCTWAS results.}

\item{anot_index}{An integer indicating how significant results are to be annotated. (-log10(TWAS.P) > anot_index) This parameter will be ignored if anno_gene is not NULL.}

\item{ceiling_ctf}{An integer indicating how significant results are to be cut by the ceiling. (-log10(TWAS.P) > ceiling_ctf)}

\item{pts_size}{An integer indicating the point size.}

\item{anno_gene}{A list of genes that need to be annotated.}

\item{path}{Path for saving the plot.}
}
\value{
A Manhattan plot
}
\description{
Manhattan Plot For TWAS Results
}
\examples{
mhp_twas(test$meta_data)
}
