% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{sim_null}
\alias{sim_null}
\title{Simulate expression z-values from null gene regulated expression distribution across tissues}
\usage{
sim_null(gene, cov_matrix, n)
}
\arguments{
\item{gene}{The gene name}

\item{cov_matrix}{The null gene regulated expression covariance matrix across tissues for all genes in the reference panel}

\item{n}{Number of simulations}
}
\value{
A nxm matrix, where n is the number of samples (simulated replicated z-values) for all expression-activated tissues, and m is the tissue number
}
\description{
Simulate expression z-values from null gene regulated expression distribution across tissues
}
\examples{
x <- sim_null("APOE", cov_matrix, 10000)
}
