% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{venn_diagram}
\alias{venn_diagram}
\title{Venn diagram for significant GReX associations}
\usage{
venn_diagram(meta_data, sctwas_res, path = NULL)
}
\arguments{
\item{meta_data}{meta_data from run_SCTWAS results.}

\item{sctwas_res}{sctwas_res from run_SCTWAS results.}

\item{path}{Path for saving the plot.}
}
\value{
A Venn diagram
}
\description{
Venn diagram for significant GReX associations
}
\examples{
venn_diagram(test$meta_data, test$sctwas_res)
}
