% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{convert_genes_ids}
\alias{convert_genes_ids}
\title{Convert Ensemble ID to gene symble in a dataframe}
\usage{
convert_genes_ids(twas)
}
\arguments{
\item{twas}{a dataframe of TWAS results with Ensemble ID}
}
\value{
a dataframe of TWAS results with gene symble
}
\description{
Convert Ensemble ID to gene symble in a dataframe
}
\examples{
twas_z = data.frame(ID = "ENSG00000261456.5", Z = "0.15066")
twas_z <= convert_genes_ids(twas_z)
}
