% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_SCTWAS}
\alias{run_SCTWAS}
\title{Run Subset-based Cross-tissue TWAS}
\usage{
run_SCTWAS(
  path,
  cov_matrix = "cov_matrix_GRCh37",
  cov_matrix_path = NULL,
  percent_act_tissue = 0.7,
  n_more = 10000,
  gene_list = NULL,
  pattern = ".alldat"
)
}
\arguments{
\item{path}{A string of the direction for TWAS results.}

\item{cov_matrix}{A string indicating list of matrix of the gene expression covariance matrix across tissues from the reference panel (default using cov_matrix_GRCh37, can also change to cov_matrix_GRCh38 or use your own matrix list). This parameter is omitted if cov_matrix_path is specified.}

\item{cov_matrix_path}{Path for downloaded reference gene expression covariance matrix across tissues (need to be named as "cov_matrix") (the reference matrix can be downloaded from: https://github.com/Thewhey-Brian/SCTWAS) If NULL, the function will automatically download the reference panel indicated by cov_matrix.}

\item{percent_act_tissue}{A decimal of the minimum percent of activated tissues for each gene regulated expression.}

\item{n_more}{Simulation times for small p-values (default 1e+04; Caution: a very large number may lead to long calculation time; a very small number may lead to inaccurate p-value estimation).}

\item{gene_list}{An array of the list of interested genes (default NULL; if NULL, it will go over all genes in the TWAS results; if not NULL, percent_act_tissue will be ignored).}

\item{pattern}{A string of the file pattern for TWAS results (default ".alldat").}
}
\value{
sctwas_res: A dataframe for the Subset-based Cross-tissue TWAS results.
meta_data: A dataframe for the tissue-specific TWAS results across multiple tissues.
}
\description{
Run Subset-based Cross-tissue TWAS
}
\examples{
res_SCTWAS = run_SCTWAS("path_to_TWAS_resutls", cov_matrix = "cov_matrix_GRCh37")
}
