% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{venn_diagram}
\alias{venn_diagram}
\title{Venn diagram for significant GReX associations}
\usage{
venn_diagram(meta_data, cstwas_res, merge_range = 1000, path = NULL)
}
\arguments{
\item{meta_data}{meta_data from run_CSTWAS results.}

\item{cstwas_res}{cstwas_res from run_CSTWAS results.}

\item{merge_range}{An integer indicating how wide (in base pairs) should be considered to merge nearby genes. Default +/- 1000bp.}

\item{path}{Path for saving the plot.}
}
\value{
A Venn diagram
}
\description{
Venn diagram for significant GReX associations
}
\examples{
venn_diagram(test$meta_data, test$cstwas_res)
}
